fix: 修复上传图片接口返回数据无法被正确识别

+++ PiliPod/Models/CommentPostModels.swift
@@ -25,7 +25,7 @@ struct CommentPictureUploadPayload: Codable {
     let imgSrc: String
     let imgWidth: Int
     let imgHeight: Int
-    let imgSize: Int
+    let imgSize: Double
 
     enum CodingKeys: String, CodingKey {
         case imgSrc = "img_src"
